Rename misago.cache.cache to misago.cache.versions
