+++ misago/cache/tests/test_getting_cache_versions.py
@@ -2,7 +2,7 @@ from unittest.mock import patch
 
 from django.test import TestCase
 
-from misago.cache.cache import (
+from misago.cache.versions import (
     CACHE_NAME, get_cache_versions, get_cache_versions_from_cache, get_cache_versions_from_db
 )
