+++ misago/cache/tests/test_invalidate_caches_management_command.py
@@ -5,7 +5,7 @@ from django.test import TestCase
 
 
 class InvalidateCachesManagementCommandTests(TestCase):
-    @patch("misago.cache.cache.invalidate_all_caches")
+    @patch("misago.cache.versions.invalidate_all_caches")
     def test_management_command_invalidates_all_caches(self, invalidate_all_caches):
         call_command('invalidateversionedcaches', stdout=Mock())
         invalidate_all_caches.assert_called_once()
